+++ src/main/resources/templates/main.html
@@ -48,4 +48,5 @@
 <script src="/js/main.js"></script>
 <script src="/js/infoWindow.js"></script>
 <script src="/js/address.js"></script>
+<script src="/js/findStore.js"></script>
 </html>
